fix a bug in the total_entry_count calculation in file_io.cpp

+++ cpp/src/wholememory/file_io.cpp
@@ -133,7 +133,7 @@ static void read_file_list_to_local_memory_roundrobin(char* local_ptr,
   size_t local_entry_file_start_index =
     local_entry_memory_start_index - memory_offset / memory_entry_stride;
 
-  size_t extra_entry       = total_file_entry_count % (wm_world_size * round_robin_size);
+  int extra_entry          = total_file_entry_count % (wm_world_size * round_robin_size);
   size_t local_extra_entry = (extra_entry > (wm_rank + 1) * round_robin_size)
                                ? round_robin_size
                                : extra_entry - wm_rank * round_robin_size;
@@ -442,7 +442,7 @@ static void read_file_list_to_local_memory_roundrobin_directio(
   size_t local_entry_file_start_index =
     local_entry_memory_start_index - memory_offset / memory_entry_stride;
 
-  size_t extra_entry       = total_file_entry_count % (wm_world_size * round_robin_size);
+  int extra_entry          = total_file_entry_count % (wm_world_size * round_robin_size);
   size_t local_extra_entry = (extra_entry > (wm_rank + 1) * round_robin_size)
                                ? round_robin_size
                                : extra_entry - wm_rank * round_robin_size;
